+++ src/client/webpack.config.js
@@ -2,6 +2,7 @@ const webpack = require("webpack");
 const path = require("path");
 const HtmlWebpackPlugin = require("html-webpack-plugin");
 const ReactRefreshPlugin = require("@pmmmwh/react-refresh-webpack-plugin");
+const MiniCssExtractPlugin = require("mini-css-extract-plugin");
 const sharedConfig = require("../webpack.shared.config.js"); // do not use `path.resolve` here
 
 const isDevelopment = sharedConfig.mode !== "production";
@@ -19,13 +20,14 @@ module.exports = {
     main: [
       // https://github.com/webpack-contrib/webpack-hot-middleware#config
       isDevelopment && "webpack-hot-middleware/client?path=/__webpack_hmr&timeout=20000&reload=true",
-      path.resolve(__dirname, "./main.tsx")
-    ].filter(Boolean)
+      path.resolve(__dirname, "./main.tsx") // @TODO rename e.g. spa.tsx ?
+    ].filter(Boolean),
+    root: path.resolve(__dirname, "./root.tsx") // @TODO rename e.g. ssr.tsx ?
   },
 
   output: {
     // path: path.resolve(__dirname, "../../build/public"),
-    // filename: "js/main.js"
+    // filename: "js/[name].js"
     path: process.env.WEBPACK_CLIENT_APP_OUTPUT_DIR,
     filename: process.env.WEBPACK_CLIENT_APP_OUTPUT_FILENAME,
     publicPath: "/"
@@ -35,6 +37,34 @@ module.exports = {
     ...sharedConfig.module,
     rules: [
       ...sharedConfig.module.rules,
+      // {
+      //   // @TODO check if webpack@5 asset modules can replace all these additional loaders
+      //   test: /\.s[ac]ss$/i,
+      //   use: [
+      //     // @NOTE NEVER use both `style-loader` and `MiniCssExtractPlugin` together
+      //     isDevelopment
+      //       ?
+      //         // Creates `style` nodes from JS strings
+      //         {
+      //           loader: "style-loader",
+      //           options: {
+      //             // this is so that it doesn't clash with React 18 SSR hydrating full-node
+      //             // as discussed here: https://github.com/facebook/react/issues/24430#issuecomment-1440427646
+      //             // @TODO consider removing this config once React solved this issue
+      //             insert: "body"
+      //           }
+      //         }
+      //       :
+      //         // strip CSS from output bundle
+      //         MiniCssExtractPlugin.loader,
+
+      //     // Translates CSS into CommonJS
+      //     "css-loader",
+
+      //     // Compiles Sass to CSS
+      //     "sass-loader"
+      //   ]
+      // },
       {
         // only for webpack@5
         // https://github.com/webpack/webpack/issues/11467#issuecomment-691873586
@@ -66,6 +96,11 @@ module.exports = {
       APP_NAME: process.env.APP_NAME || ""
     }),
     new HtmlWebpackPlugin({
+
+      // the JS bundle from the `main` entry used for the SPA mode;
+      // note that we don't care about the `root` entry here (which is used for a completely different purpose)
+      chunks: ['main'],
+
       // @NOTE inject build-time vars as needed
       filename: isDevelopment
 
@@ -87,7 +122,16 @@ module.exports = {
       overlay: {
         sockIntegration: "whm",
       }
-    })
+    }),
+    // // this is here (instead of `webpack.shared.config`) because of the way we include & configure it
+    // new MiniCssExtractPlugin({
+    //   // Options similar to the same options in webpackOptions.output
+    //   // all options are optional
+    //   filename: "css/[name].css",
+    //   chunkFilename: "css/[id].css",
+    //   // Enable to remove warnings about conflicting order
+    //   // ignoreOrder: false,
+    // }),
   ].filter(Boolean)
 
   /* the following optimization is disabled for the sake of simplicity for the exercise */