+++ src/server/global.d.ts
@@ -1,0 +1,2 @@
+// @TODO remove this file if proven unnecessary / unused
+// declare module '*.scss';